Finish splitting guiGate.cpp

+++ src/gui/command/cmdConnectWire.cpp
@@ -118,7 +118,7 @@ void cmdConnectWire::sendMessagesToConnect(GUICircuit *gCircuit, IDType wireId,
 		internalHotspots.push_back(hotspot);
 	}
 	else {
-		for (int i = 0; i < wireIds.size(); i++) {
+		for (int i = 0; i < (int)wireIds.size(); i++) {
 			internalHotspots.push_back(hotspot + "_" + std::to_string(i));
 		}
 	}
@@ -126,7 +126,7 @@ void cmdConnectWire::sendMessagesToConnect(GUICircuit *gCircuit, IDType wireId,
 	bool isInput = gate->isConnectionInput(hotspot);
 
 	// Connect each of wire's bus-lines to its corresponding gate hotspot.
-	for (int i = 0; i < internalHotspots.size(); i++) {
+	for (int i = 0; i < (int)internalHotspots.size(); i++) {
 		if (isInput) {
 			gCircuit->sendMessageToCore(
 				new Message_SET_GATE_INPUT(gateId, internalHotspots[i], wireIds[i]));
@@ -160,7 +160,7 @@ void cmdConnectWire::sendMessagesToDisconnect(GUICircuit *gCircuit,
 		internalHotspots.push_back(hotspot);
 	}
 	else {
-		for (int i = 0; i < wireIds.size(); i++) {
+		for (int i = 0; i < (int)wireIds.size(); i++) {
 			internalHotspots.push_back(hotspot + "_" + to_string(i));
 		}
 	}
@@ -168,7 +168,7 @@ void cmdConnectWire::sendMessagesToDisconnect(GUICircuit *gCircuit,
 	bool isInput = gate->isConnectionInput(hotspot);
 
 	// Disconnect each of wire's bus-lines from its corresponding gate hotspot.
-	for (int i = 0; i < internalHotspots.size(); i++) {
+	for (int i = 0; i < (int)internalHotspots.size(); i++) {
 		if (isInput) {
 			gCircuit->sendMessageToCore(
 				new Message_SET_GATE_INPUT(gateId, internalHotspots[i], 0, true));
